build list
